tests: Add QS SyncVal disable support and test
QueueSubmit validation is now on by default.  Add disable option and
test for the disable actually suppressing validation.
