+++ tests/CMakeLists.txt
@@ -52,6 +52,8 @@ target_sources(vk_layer_validation_tests PRIVATE
     framework/gpu_av_helper.h
     framework/render_pass_helper.h
     framework/render_pass_helper.cpp
+    framework/queue_submit_context.h
+    framework/queue_submit_context.cpp
     unit/amd_best_practices.cpp
     unit/android_hardware_buffer.cpp
     unit/android_hardware_buffer_positive.cpp
