+++ tests/framework/layer_validation_tests.h
@@ -290,7 +290,7 @@ class NegativeDebugPrintf : public VkLayerTest {
 
 class VkSyncValTest : public VkLayerTest {
   public:
-    void InitSyncValFramework(bool enable_queue_submit_validation = false);
+    void InitSyncValFramework(bool disable_queue_submit_validation = false);
 
   protected:
     const VkValidationFeatureEnableEXT enables_[1] = {VK_VALIDATION_FEATURE_ENABLE_SYNCHRONIZATION_VALIDATION_EXT};
